refactor: search 및 watch 디렉토리 구조 병합
