+++ src/js/youtube-class-room/watch/WatchController.js
@@ -1,8 +1,8 @@
 import WatchView from "./WatchView.js";
 
-import { CONFIRM_MESSAGE } from "../utils/constants.js";
+import { CONFIRM_MESSAGE } from "../../utils/constants.js";
 
-import videos from "../state/videos.js";
+import videos from "../../state/videos.js";
 export default class WatchController {
   constructor() {
     this.watchView = new WatchView();
